Move Python binding of annoy to another file.
Fixes #5.
