+++ setup.py
@@ -33,7 +33,7 @@ setup(name='annoy',
       version='1.0.3',
       description='Approximate Nearest Neighbors in C++/Python optimized for memory usage and loading/saving to disk.',
       packages=['annoy'],
-      ext_modules=[Extension('annoy.annoylib', ['src/annoylib.cc'], libraries=['boost_python'])],
+      ext_modules=[Extension('annoy.annoylib', ['src/annoymodule.cc'], libraries=['boost_python'])],
       long_description=long_description,
       author='Erik Bernhardsson',
       author_email='erikbern@spotify.com',
