+++ src/annoylib.h
@@ -12,10 +12,11 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-#include "Python.h"
+#ifndef ANNOYLIB_H
+#define ANNOYLIB_H
+
 #include <stdio.h>
 #include <string>
-#include <boost/python.hpp>
 #include <sys/stat.h>
 #include <unistd.h>
 #include <stdio.h>
@@ -519,63 +520,4 @@ protected:
   }
 };
 
-template<typename T, typename Distance>
-class AnnoyIndexPython : public AnnoyIndex<T, Distance > {
-public:
-  AnnoyIndexPython(int f): AnnoyIndex<T, Distance>(f) {}
-  void add_item_py(int item, const python::list& v) {
-    vector<T> w;
-    for (int z = 0; z < this->_f; z++)
-      w.push_back(python::extract<T>(v[z]));
-
-    this->add_item(item, &w[0]);
-  }
-  python::list get_nns_by_item_py(int item, size_t n) {
-    vector<int> result;
-    this->get_nns_by_item(item, n, &result);
-    python::list l;
-    for (size_t i = 0; i < result.size(); i++)
-      l.append(result[i]);
-    return l;
-  }
-  python::list get_nns_by_vector_py(python::list v, size_t n) {
-    vector<T> w(this->_f);
-    for (int z = 0; z < this->_f; z++)
-      w[z] = python::extract<T>(v[z]);
-    vector<int> result;
-    this->get_nns_by_vector(&w[0], n, &result);
-    python::list l;
-    for (size_t i = 0; i < result.size(); i++)
-      l.append(result[i]);
-    return l;
-  }
-  python::list get_item_vector_py(int item) {
-    const typename Distance::node* m = this->_get(item);
-    const T* v = m->v;
-    python::list l;
-    for (int z = 0; z < this->_f; z++) {
-      l.append(v[z]);
-    }
-    return l;
-  }
-};
-
-template<typename C>
-void expose_methods(python::class_<C> c) {
-  c.def("add_item",          &C::add_item_py)
-    .def("build",             &C::build)
-    .def("save",              &C::save)
-    .def("load",              &C::load)
-    .def("unload",            &C::unload)
-    .def("get_distance",      &C::get_distance)
-    .def("get_nns_by_item",   &C::get_nns_by_item_py)
-    .def("get_nns_by_vector", &C::get_nns_by_vector_py)
-    .def("get_item_vector",   &C::get_item_vector_py)
-    .def("get_n_items",       &C::get_n_items);
-}
-
-BOOST_PYTHON_MODULE(annoylib)
-{
-  expose_methods(python::class_<AnnoyIndexPython<float, Angular<float> > >("AnnoyIndexAngular", python::init<int>()));
-  expose_methods(python::class_<AnnoyIndexPython<float, Euclidean<float> > >("AnnoyIndexEuclidean", python::init<int>()));
-}
+#endif
